+++ src/layers/VivViewerLayer/VivViewerLayerBase.js
@@ -1,8 +1,6 @@
 import { TileLayer } from '@deck.gl/geo-layers';
 import { COORDINATE_SYSTEM } from '@deck.gl/core';
-
-import { loadZarr, loadTiff } from './data-utils';
-import { isInTileBounds, renderSubLayers } from './utils';
+import { renderSubLayers } from './utils';
 
 const defaultProps = {
   ...TileLayer.defaultProps,
@@ -15,39 +13,11 @@ const defaultProps = {
   imageWidth: { type: 'number', value: 1028, compare: true },
   minZoom: { type: 'number', value: 0, compare: true },
   maxZoom: { type: 'number', value: 0, compare: true },
-  connections: { type: 'object', value: {}, compare: true },
-  pool: { type: 'object', value: {}, compare: false },
   renderSubLayers: { type: 'function', value: renderSubLayers, compare: false },
-  useZarr: { type: 'boolean', value: false, compare: true },
-  useTiff: { type: 'boolean', value: false, compare: true },
   opacity: { type: 'number', value: 1, compare: true }
 };
 
-export default class VivViewerLayerBase extends TileLayer {
-  async getTileData({ x, y, z }) {
-    const { useZarr, useTiff } = this.props;
-    const tileLoadProps = {
-      x,
-      y,
-      z: -z,
-      ...this.props
-    };
-    if (isInTileBounds(tileLoadProps)) {
-      if (useTiff) {
-        const tile = loadTiff(tileLoadProps);
-        return tile;
-      }
-      if (useZarr) {
-        const tile = loadZarr(tileLoadProps);
-        return tile;
-      }
-      // If neither zarr or tiff is indicated, use whatever is passed in.
-      const tile = this.props.getTileData({ x, y, z });
-      return tile;
-    }
-    return null;
-  }
-}
+export default class VivViewerLayerBase extends TileLayer {}
 
 VivViewerLayerBase.layerName = 'VivViewerLayerBase';
 VivViewerLayerBase.defaultProps = defaultProps;