+++ src/layers/VivViewerLayer/data-utils/index.js
@@ -1,4 +1,0 @@
-import { loadZarr, initZarr } from './zarr-utils';
-import { loadTiff, initTiff } from './tiff-utils';
-
-export { loadZarr, loadTiff, initZarr, initTiff };